requirements.txt
pandas
pyarrow
langchain
langchain-google-genai
python-dotenv